pyproject.toml
[project]
name = "sfra-full"
version = "0.1.0-phase0"
description = "APTRANSCO SFRA analysis platform — IEEE C57.149-2012 compliant per-combination diagnostics with versioned reference library."
readme = "external/INVENTORY.md"
requires-python = ">=3.11"
authors = [
    { name = "APTRANSCO HIS / SFRA platform team" },
]
license = { text = "Proprietary — APTRANSCO internal use" }
keywords = [
    "sfra",
    "transformer",
    "ieee-c57-149",
    "iec-60076-18",
    "cigre",
    "dl-t-911",
    "aptransco",
]

# Runtime dependencies. Pinned to known-good versions that align with the
# upstream external/SFRA dependency set so we can import its primitives
# without resolution conflicts.
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.32.0",
    "python-multipart==0.0.12",
    "pydantic==2.9.2",
    "sqlalchemy>=2.0.30,<2.1",
    "alembic>=1.13,<2.0",
    # Spec v2 §2: PostgreSQL is the production DB. SQLite is only for local
    # dev. psycopg[binary] is the v3 driver; falls back to plain psycopg
    # when system libpq is available.
    "psycopg[binary]>=3.2,<4",
    "numpy==1.26.4",
    "scipy==1.13.1",
    "pandas==2.2.2",
    "matplotlib>=3.8,<3.10",       # headless PDF rendering
    "plotly>=5.22,<6",             # interactive plot JSON for the frontend
    "reportlab==4.2.2",
    "openpyxl>=3.1,<4",            # XLSX export
    "lxml==5.3.0",                 # FRAX XML parsing (spec v2 §4.1)
    "pyyaml>=6.0,<7",
    "python-jose[cryptography]>=3.3,<4",  # JWT for Engineer/Reviewer/Admin (spec v2 §2)
    "passlib[bcrypt]>=1.7,<2",
    "boto3>=1.34,<2",              # pluggable S3 storage backend (spec v2 §2)
    "httpx==0.27.2",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.3",
    "pytest-cov>=5.0,<7",
    "pytest-asyncio>=0.23,<1",
    "ruff>=0.6,<1",
    "black>=24.4,<26",
    "mypy>=1.10,<2",
    "types-PyYAML",
    "pre-commit>=3.7,<4",
    "ipython>=8.20",
]

[project.scripts]
sfra-full = "sfra_full.cli:main"

[project.urls]
"Standards reference (IEEE)" = "https://standards.ieee.org/ieee/C57.149/4528/"
"APTRANSCO"                  = "https://aptransco.gov.in/"

# ---------------------------------------------------------------------------
# Build backend
# ---------------------------------------------------------------------------
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sfra_full"]

# ---------------------------------------------------------------------------
# uv configuration
# ---------------------------------------------------------------------------
[tool.uv]
# Allow editable installs from the upstream external/SFRA so we can import
# its analysis primitives in-place. Developers must run scripts/setup_external.sh
# first; until then this source is unavailable and the relevant imports raise
# at runtime with a clear message (see src/sfra_full/external_bridge.py).
dev-dependencies = []

# ---------------------------------------------------------------------------
# Ruff (lint + import sort)
# ---------------------------------------------------------------------------
[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]
extend-exclude = ["external", "frontend"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
    "N",   # pep8-naming
    "RUF", # ruff-specific
]
ignore = [
    "E501", # line length handled by black
]

[tool.ruff.lint.isort]
known-first-party = ["sfra_full"]

# ---------------------------------------------------------------------------
# Black
# ---------------------------------------------------------------------------
[tool.black]
line-length = 100
target-version = ["py311"]
extend-exclude = "external|frontend"

# ---------------------------------------------------------------------------
# mypy — strict on analysis modules per spec §5
# ---------------------------------------------------------------------------
[tool.mypy]
python_version = "3.11"
files = ["src/sfra_full"]
strict_optional = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unreachable = true
disallow_untyped_decorators = true

[[tool.mypy.overrides]]
module = "sfra_full.analysis.*"
strict = true

[[tool.mypy.overrides]]
module = ["scipy.*", "skrf.*", "matplotlib.*", "plotly.*", "reportlab.*", "openpyxl.*", "lxml.*"]
ignore_missing_imports = true

# ---------------------------------------------------------------------------
# Pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--cov=sfra_full",
    "--cov-report=term-missing",
    "--cov-fail-under=0",  # Phase 0 baseline; raised to 80 in Phase 1 per spec §5
]
markers = [
    "slow: tests that take more than ~1s",
    "integration: tests that hit the database or filesystem",
    "needs_external: tests that require external/SFRA to be cloned",
]

[tool.coverage.run]
branch = true
source = ["src/sfra_full"]
omit = ["src/sfra_full/cli.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
    "@abstractmethod",
]
